Changed how Instructions content is loaded
old behavior: instructions page was loaded and cached so that changes
to the web page were not reflected in the app.

new behavior: everytime the app starts it re-loads the instructions
content. this is good for testing/modifying the instructions content
ahead of release. likely not necessary once released. causes a hit to
startup time.