Enable/Disable Save button based on changes to user settings
Settings pane opens w/ Save button disabled. It will only be enabled if
the user changes the values of any of the properties. Tapping save will
save that data (as it does now) and then disable the button again.

This is to provide some subtle feedback that something needs to be
saved or not.